LayoutTests/http/wpt/navigation-api/transition-promises-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error
CONSOLE MESSAGE: Unhandled Promise Rejection: Error

Harness Error (FAIL), message = Unhandled rejection

PASS NavigationTransition finished promise is fulfilled on success
PASS NavigationTransition finished promise is rejected on error
